Improve ReorderableWrap for full width widgets
Changes the drag targets from two horizontal/vertical rectangles to two triangles covering the upper left side for the "preDragTarget" and the lower right side for the "nextDragTarget".
When a reorderable widget expands to full width, it is now more intuitive to drag it to the previous/next position.

For example, without this change, the user always had to drag the widget to the left to move it to the previous position when horizontal mode was enabled. Now it is possible to simply drag it upwards on the right side of the widget.